+++ test/notifications.test.js
@@ -10,12 +10,7 @@ const {
 describe('notification sending', () => {
   const settingsSendAlerts = {
     communications: {
-      send_alerts: true,
-    },
-  };
-  const settingsDontSendAlerts = {
-    communications: {
-      send_alerts: false,
+      channel: 1,
     },
   };
   it('should rank EMERGENCY as sendable', () => {
@@ -30,19 +25,6 @@ describe('notification sending', () => {
       settingsSendAlerts,
     );
     assert.equal(result, true);
-  });
-  it('should rank EMERGENCY as not sendable if alert sending is disabled', () => {
-    const episodes = new Map();
-    const result = shouldWeSendNotification(
-      'notifications.communication.meshtastic.deviceStateNum',
-      {
-        state: 'emergency',
-        message: 'Disconnected from Meshtastic node',
-      },
-      episodes,
-      settingsDontSendAlerts,
-    );
-    assert.equal(result, false);
   });
   it('should rank NOMINAL as not sendable', () => {
     const episodes = new Map();